refactor(GameIdentifier): factor out versionInfo constructor for JenkinsRepositoryAdapter
