+++ src/main/java/org/terasology/launcher/repositories/JenkinsRepositoryAdapter.java
@@ -3,7 +3,6 @@
 
 package org.terasology.launcher.repositories;
 
-import com.vdurmont.semver4j.Semver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.terasology.launcher.model.Build;
@@ -97,23 +96,13 @@ class JenkinsRepositoryAdapter implements ReleaseRepository {
     private Optional<GameIdentifier> computeIdentifierFrom(Jenkins.Build jenkinsBuildInfo) {
         return Optional.ofNullable(client.getArtifactUrl(jenkinsBuildInfo, "versionInfo.properties"))
                 .map(client::requestProperties)
-                .map(versionInfo -> {
-                    var displayVersion = versionInfo.getProperty("displayVersion");
-                    if (displayVersion == null) {
-                        return null;
-                    }
+                .flatMap(versionInfo -> GameIdentifier.fromVersionInfo(versionInfo, buildProfile, profile))
+                .map(id -> {
                     // The Jenkins adapter has to append the build number to the display version to ensure that the
                     // version string is unique for two different builds. This is necessary because the display version
                     // generated by the CI build is the same of subsequent builds...
-                    String buildVersion = displayVersion + "+" + jenkinsBuildInfo.number;
-                    try {
-                        Semver engineVersion = new Semver(versionInfo.getProperty("engineVersion"), Semver.SemverType.IVY);
-                        return new GameIdentifier(buildVersion, engineVersion, buildProfile, profile);
-                    } catch (RuntimeException e) {
-                        logger.error("versionInfo.properties displayVersion=\"{}\" engineVersion=\"{}\" " +
-                                "is not a valid version.", displayVersion, versionInfo.getProperty("engineVersion"), e);
-                        return null;
-                    }
+                    return new GameIdentifier(id.getVersion() + "+" + jenkinsBuildInfo.number,
+                            id.getEngineVersion(), id.getBuild(), id.getProfile());
                 });
     }
